refactor: 필요 없는 메서드 삭제
