+++ src/main/kotlin/blackjack/business/participants/Players.kt
@@ -1,8 +1,5 @@
 package blackjack.business.participants
 
-import blackjack.business.card.CardDesk
-import blackjack.business.drawConditionStrategy.DrawConditionStrategy
-
 class Players private constructor(allPlayers: List<Player>) {
 
     private val _players: List<Player> = allPlayers
@@ -10,26 +7,6 @@ class Players private constructor(allPlayers: List<Player>) {
     val allPlayers: List<Player> = _players.toList()
 
     fun forEachPlayer(onPlayerAction: (Player) -> Unit) = allPlayers.forEach(onPlayerAction)
-
-    fun dealInitialCards(cardDesk: CardDesk, onPlayerAction: (Player) -> Unit) {
-        forEachPlayer { player ->
-            repeat(2) { player.addCard(cardDesk.draw()) }
-            onPlayerAction(player)
-        }
-    }
-
-    fun processAdditionalCards(
-        cardDesk: CardDesk,
-        drawConditionStrategy: DrawConditionStrategy,
-        onPlayerAction: (Player) -> Unit,
-    ) {
-        forEachPlayer { player ->
-            while (player.canDrawCard() && drawConditionStrategy.shouldDraw(player.name)) {
-                player.addCard(cardDesk.draw())
-                onPlayerAction(player)
-            }
-        }
-    }
 
     init {
         require(allPlayers.size > 1) { "플레이어는 2명 이상이여야 가능합니다." }
